Migrate data package to an OpenSpending Data Package profile
OpenSpending Data Package is a proposed data package profile for
OpenSpending specifically, based on Budget Data Packages but uses
a mapping to provide mapping between source files and attributes
instead of enforcing specific headers on the source files.

This is a working version but it's missing quite a lot of mapping
as the current draft of the OpenSpending Data Package does not
provide mapping infrastructure which suits this datapackage well.